xatu: size the count fields so the page cache can use SSZ
The arrival models used plain int for observation and node counts, which
SSZ cannot encode, so every cache write for these pages fell back to JSON
and logged a compatibility warning. A mainnet-scale slot payload is a few
hundred kilobytes, so that fallback is not free.

Counts are non-negative and bounded, so uint32 is the honest type and the
cache now takes the SSZ path.

Also adds a regression test for the payload size itself: MaxEntrySize is
100KB, which reads like a limit these payloads would breach, but it only
sizes the initial allocation and the real ceiling is HardMaxCacheSize per
shard. Worth pinning, because if large entries ever stop being cached the
symptom is silent load rather than an error.

+++ handlers/epoch_arrival.go
@@ -185,7 +185,7 @@ func buildEpochArrivalData(ctx context.Context, epoch phase0.Epoch) (*models.Epo
 	for slot, values := range slotValues {
 		sort.Slice(values, func(a, b int) bool { return values[a] < values[b] })
 		response.Slots[uint64(slot)] = &models.EpochArrivalSlot{
-			Nodes: len(values),
+			Nodes: uint32(len(values)),
 			MinMs: values[0],
 			P90Ms: values[len(values)*9/10],
 		}
